Added checks for invalid filenames. (SD2-394)
git-svn-id: svn://svn.sharpdevelop.net/sharpdevelop/trunk@307 1ccf3a8d-04fe-1044-b7c0-cef0b8235c61

+++ src/Main/Base/Project/Src/Gui/Pads/ProjectBrowser/TreeNodes/DirectoryNode.cs
@@ -124,7 +124,7 @@ namespace ICSharpCode.SharpDevelop.Project
 				}
 			}
 		}
-		public override void Refresh() 
+		public override void Refresh()
 		{
 			base.Refresh();
 			if (Nodes.Count == 0) {
@@ -352,13 +352,16 @@ namespace ICSharpCode.SharpDevelop.Project
 			if (newName == null) {
 				return;
 			}
+			if (!FileService.CheckFileName(newName)) {
+				return;
+			}
 			string oldText = Text;
 			Text = newName;
 			if (Directory != null) {
 				string newPath = Path.Combine(Path.GetDirectoryName(Directory), newName);
 				if (System.IO.Directory.Exists(newPath)) {
 					if (System.IO.Directory.GetFiles(Directory).Length == 0) {
-						System.IO.Directory.Delete(Directory); 
+						System.IO.Directory.Delete(Directory);
 					} else if (System.IO.Directory.GetFiles(newPath).Length == 0) {
 						System.IO.Directory.Delete(newPath);
 						FileService.RenameFile(Directory, newPath, true);
